resources/python_cloud_functions/configurator/requirements.txt
attrs==22.1.0
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-api-core==2.10.1
google-auth==2.11.1
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.6
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
